requirements.txt
langchain==0.1.16
llama-index==0.9.48
farm-haystack==1.26.4
openai==1.13.3
transformers==4.46.1  # Upgraded to match Farm-Haystack requirements
opencv-python==4.8.0.74
pillow==9.5.0
sqlalchemy==2.0.15
fastapi==0.110.0
uvicorn==0.24.0
loguru==0.7.2
sentry-sdk==1.5.0
python-dotenv==1.0.0
pydantic==1.10.12  # Compatible with LangChain & Farm-Haystack
typing-extensions==4.8.0  # Compatible with FastAPI & OpenAI
